+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Changed
+- The Summary section of the report now only lists the first 7 sample names and reference files, instead of listing them all.
+
 ## [v1.1.1]
 ### Changed
 - Reduced the memory requested by some processes to avoid failing in WSL (since there is slightly less memory available in WSL than specified in `.wslconfig`).